requirements.txt
# requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.0
Pillow==10.0.0
folium==0.15.0
streamlit-folium==0.15.0
scikit-learn==1.4.0